requirements.txt
discord.py
spotipy
google-api-python-client